requirements.txt
dash==2.17.1
pandas==2.2.2
numpy==2.0.0
dash-bootstrap-components==1.6.0
alembic==1.13.2
pyyaml==6.0.1
python-dotenv==1.0.1
fastapi==0.112.0
pyjwt==2.9.0
flask-login==2.1.5
sqlalchemy==3.0.3
aiosqlite==0.20.0
streaming_form_data==1.16.0
iso8601==2.1.0
./modules/dist/python_modules-0.1-py3-none-any.whl
